# lib/api-zod/src/generated/types/chatSessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Support Connect - Live WhatsApp Support Chat App
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSessionInput {
  userName: string;
  userPhone?: string;
}